C12: ft_sorted_list_merge; refactor cmake

+++ c00/src/tests/ft_print_comb_test.c
@@ -66,18 +66,18 @@ int ft_print_comb_test(void)
 
 	bytes_read = run_on_pipe(buf);
 
-	check_val = (bytes_read == ft_strlen(checkbuf));
+	check_val = (bytes_read == (ssize_t)ft_strlen(checkbuf));
 	if (!check_val)
 	{
 		fprintf(stdout, "bytes_read doesn't match expected value\n");
-		fprintf(stdout, "got return size \"%d\" whilst [\"%d\"] was to be expected\n",
+		fprintf(stdout, "got return size \"%zd\" whilst [\"%zu\"] was to be expected\n",
 				bytes_read, ft_strlen(checkbuf));
 	}
 	if (bytes_read > 0)
 	{
 		check_val = !ft_strcmp(buf, checkbuf);
 		if (!check_val)
-			fprintf(stdout, "got \"%.*s\" whilst \"%s\" was to be expected\n", bytes_read, buf, checkbuf);
+			fprintf(stdout, "got \"%.*s\" whilst \"%s\" was to be expected\n", (int)bytes_read, buf, checkbuf);
 	}
 	check(check_val);
 	return (0);
